3.4-3.7 & Indiv1

+++ Lab3.4/Lab3.4.cs
@@ -10,9 +10,9 @@ namespace Lab3._4
 	{
 		static void Main(string[] args)
 		{
-			double[,] Sum(double[,] a, double[,] b, double[,] avg)
+			float[,] Sum(float[,] a, float[,] b, float[,] avg)
 			{
-				double[,] ans = new double[3,3];
+				float[,] ans = new float[3,3];
 				for (int i = 0; i < 3; i++)
 					for (int j = 0; j < 3; j++)
 					{
@@ -22,9 +22,9 @@ namespace Lab3._4
 				return ans;
 			}
 
-			double[,] Diff(double[,] a, double[,] b, double[,] avg)
+			float[,] Diff(float[,] a, float[,] b, float[,] avg)
 			{
-				double[,] ans = new double[3, 3];
+				float[,] ans = new float[3, 3];
 				for (int i = 0; i < 3; i++)
 					for (int j = 0; j < 3; j++)
 					{
@@ -35,15 +35,15 @@ namespace Lab3._4
 			}
 
 			Random rand = new Random();
-			double[,] x = new double[3, 3];
-			double[,] y = new double[3, 3];
-			double[,] z = new double[3, 3];
-			double[,] aver = new double[3, 3];
+			float[,] x = new float[3, 3];
+			float[,] y = new float[3, 3];
+			float[,] z = new float[3, 3];
+			float[,] aver = new float[3, 3];
 			for (byte i = 0; i < 3; i++)
 				for (byte j = 0; j < 3; j++)
 				{
-					x[i, j] = (double)rand.Next(1000) / 10;
-					y[i, j] = (double)rand.Next(1000) / 10;
+					x[i, j] = (float)rand.Next(1000) / 10;
+					y[i, j] = (float)rand.Next(1000) / 10;
 				}
 
 			Console.WriteLine("Эта программа складывает или вычитает, а также вычисляет среднее значение элементов двух масивов 3х3\n\n1й Массив:\n");
